SLVSCODE-1031 Remove node-fetch (now part of Node)
